Clasificador__cabify_api/requirements.txt
fastapi
uvicorn[standard]
requests
psycopg2-binary
pydantic
python-dotenv
